Text based version put into scr--TextBasedVersion
