--- scr/Obstacle1.java
@@ -1,23 +1,0 @@
-
-public class Obstacle1 {
-	
-
-	
-	int row;
-	int column;
-	
-	public Obstacle1(int row, int column) {
-		this.row = row;
-		this.column = column;
-	}
-	
-	public int getRow() {
-		return row;
-	}
-	
-	public int getColumn() {
-		return column;
-	}
-	
-
-}
