+++ scr/TextBasedVersion/Obstacle1.java
@@ -1,0 +1,24 @@
+package CrazyCodingClub.TextBasedVersion;
+
+public class Obstacle1 {
+	
+
+	
+	int row;
+	int column;
+	
+	public Obstacle1(int row, int column) {
+		this.row = row;
+		this.column = column;
+	}
+	
+	public int getRow() {
+		return row;
+	}
+	
+	public int getColumn() {
+		return column;
+	}
+	
+
+}
